Allow schema types with the same name from different packages
Previously the short type name was used as a schema key. This meant that
types of the same name from different packages would overwrite each
other in the map of schema type definitions, causing unexpected
serialization errors. Now the fully qualified (package + type) name is
used so that entries for types with the same name from different
packages are unique.

Signed-off-by: Mark S. Lewis <Mark.S.Lewis@outlook.com>