Support struct arguments with all optional fields
Transaction functions with struct argument that have all their fields as
JSON optional causes a parsing error on the chaincode metadata:

    components.schemas.Args.required: Array must have at least 1 items

Having no required fields causes the `required` property to be an empty
array. Draft 4 of the JSON schema specification requires string arrays
to contain at least one element, so an error occurs.

This change modifies the JSON description of the ObjectMetata `Required`
property so that it is omitted when empty.

Signed-off-by: Mark S. Lewis <Mark.S.Lewis@outlook.com>